# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Image to Website API",
    "description": "Production backend for turning a single product photograph into a complete, working online store with deterministic Jinja rendering, AI vision intelligence, and server-managed commerce.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/shop/{website_id}": {
      "get": {
        "tags": [
          "Commerce & Orders"
        ],
        "summary": "Shop Dashboard",
        "description": "The merchant's order desk. Served as a plain page; it talks to the API.",
        "operationId": "shop_dashboard_shop__website_id__get",
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/{website_id}/shop": {
      "get": {
        "tags": [
          "Commerce & Orders"
        ],
        "summary": "Shop Info",
        "description": "Everything the storefront needs to render a cart and a checkout.",
        "operationId": "shop_info_api__website_id__shop_get",
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/{website_id}/quote": {
      "post": {
        "tags": [
          "Commerce & Orders"
        ],
        "summary": "Shop Quote",
        "description": "Price a cart without placing an order.\n\nThe cart page calls this instead of adding its numbers up locally, so what\nthe buyer is shown is the same figure the server will charge.",
        "operationId": "shop_quote_api__website_id__quote_post",
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/{website_id}/orders": {
      "post": {
        "tags": [
          "Commerce & Orders"
        ],
        "summary": "Place Order",
        "description": "Place an order. Public, rate limited, server-priced.",
        "operationId": "place_order_api__website_id__orders_post",
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Commerce & Orders"
        ],
        "summary": "Merchant Orders",
        "operationId": "merchant_orders_api__website_id__orders_get",
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/{website_id}/orders/{order_id}/paid": {
      "post": {
        "tags": [
          "Commerce & Orders"
        ],
        "summary": "Claim Paid",
        "description": "The buyer says they have paid and gives a reference.\n\nMoves the order to `pending`, never `paid` \u2014 a UPI deep link tells us\nnothing, so only the seller checking their bank can settle it.",
        "operationId": "claim_paid_api__website_id__orders__order_id__paid_post",
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          },
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/{website_id}/orders/{order_id}": {
      "patch": {
        "tags": [
          "Commerce & Orders"
        ],
        "summary": "Update Order",
        "operationId": "update_order_api__website_id__orders__order_id__patch",
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          },
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/{website_id}/price": {
      "post": {
        "tags": [
          "Commerce & Orders"
        ],
        "summary": "Set Item Price",
        "description": "Set what the product costs. Turns an enquiry-only site into a shop.",
        "operationId": "set_item_price_api__website_id__price_post",
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/{website_id}/settings": {
      "post": {
        "tags": [
          "Commerce & Orders"
        ],
        "summary": "Shop Settings",
        "description": "Seller settings \u2014 UPI ID, payee name, where to send new-order alerts.",
        "operationId": "shop_settings_api__website_id__settings_post",
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}